refactor: Flatten include/lua_zephyr/ and src/lua_zephyr/ into include/ and src/
Remove redundant lua_zephyr/ subdirectory nesting. Files already have
luaz_* prefix for namespacing. Update all #include paths to drop the
lua_zephyr/ prefix, update CMakeLists.txt source paths, justfile format
globs, and CLAUDE.md directory references.

All 5 twister test configurations pass.
