+++ samples/heavy/src/main.c
@@ -7,7 +7,7 @@
 
 #include <lauxlib.h>
 #include <lualib.h>
-#include <lua_zephyr/luaz_utils.h>
+#include <luaz_utils.h>
 
 /** @brief Setup hook for the heavy Lua thread (loads string + zephyr libs). */
 int heavy_lua_setup(lua_State *L)
