+++ samples/producer_consumer/src/producer_lua.c
@@ -11,8 +11,8 @@
 #include <lualib.h>
 #include <sys/times.h>
 #include <zephyr/zbus/zbus.h>
-#include <lua_zephyr/luaz_utils.h>
-#include <lua_zephyr/luaz_zbus.h>
+#include <luaz_utils.h>
+#include <luaz_zbus.h>
 
 #include "channels.h"
