+++ src/luaz_repl.c
@@ -12,7 +12,7 @@
 #include <lua.h>
 #include <lualib.h>
 #include <lauxlib.h>
-#include <lua_zephyr/luaz_utils.h>
+#include <luaz_utils.h>
 #include <zephyr/kernel.h>
 #include <zephyr/shell/shell.h>
 #include <zephyr/sys/printk.h>
