+++ templates/lua_thread.c.in
@@ -13,7 +13,7 @@
 #include <lua.h>
 #include <lauxlib.h>
 #include <lualib.h>
-#include <lua_zephyr/luaz_utils.h>
+#include <luaz_utils.h>
 #include <zephyr/kernel.h>
 
 static char heap_mem[CONFIG_LUA_THREAD_HEAP_SIZE];
